Make `main` cleaner
