+++ tools/diagnose/src/main.rs
@@ -1,7 +1,6 @@
 use clap::Parser;
 use cmd::{Commands, Format, Opts};
 use stable_eyre::{eyre::Context, Result};
-use tracing::{debug, error, info, trace, warn};
 use tracing_subscriber::{prelude::*, Layer, Registry};
 
 mod cmd;
@@ -35,63 +34,6 @@ fn main() -> Result<()> {
 
     match opts.command() {
         Commands::Walk(args) => cmd::walk::main(root, args),
-        Commands::DebugOutputFormat => debug_output_format(),
+        Commands::DebugOutputFormat => cmd::debug_output_format::main(),
     }
-}
-
-fn debug_output_format() -> Result<()> {
-    #[tracing::instrument]
-    fn inner(value: &str) {
-        error!("message at level 'Error' inside function 'inner'");
-        info!("message at level 'Info' inside function 'inner'");
-        warn!("message at level 'Warn' inside function 'inner'");
-        debug!("message at level 'Debug' inside function 'inner'");
-        trace!("message at level 'Trace' inside function 'inner'");
-        let _ = value;
-        nested_inner("some value passed to 'nested_inner'")
-    }
-
-    #[tracing::instrument]
-    fn nested_inner(value: &str) {
-        error!("message at level 'Error' inside function 'nested_inner_function', called from inside 'inner'");
-        info!("message at level 'Info' inside function 'nested_inner_function', called from inside 'inner'");
-        warn!("message at level 'Warn' inside function 'nested_inner_function', called from inside 'inner'");
-        debug!("message at level 'Debug' inside function 'nested_inner_function', called from inside 'inner'");
-        trace!("message at level 'Trace' inside function 'nested_inner_function', called from inside 'inner'");
-        let _ = value;
-    }
-
-    #[tracing::instrument(fields(uppercased))]
-    fn uppercase(input: &str) -> String {
-        let result = input.to_uppercase();
-        tracing::Span::current().record("uppercased", &result);
-        result
-    }
-
-    info!("demonstrating how messages are logged.");
-    info!("");
-    error!("message at level 'Error'");
-    info!("message at level 'Info'");
-    warn!("message at level 'Warn'");
-    debug!("message at level 'Debug'");
-    trace!("message at level 'Trace'");
-
-    let target = "hello";
-    info!("");
-    info!("-------");
-    info!("demonstrating how spans are logged.");
-    info!("calling 'uppercase' with the argument '{target}', then logging the result:");
-    info!("");
-    info!("the word '{target}' uppercased is...");
-    let result = uppercase(target);
-    info!("> '{result}'");
-
-    info!("");
-    info!("-------");
-    info!("demonstrating how nested spans are logged.");
-    info!("calling the function 'inner' with an argument, which will in turn call 'nested_inner' with a different argument.");
-    info!("");
-    inner("some value passed to 'inner'");
-
-    Ok(())
 }
